move to ObjectDumping namespace, move auxilary classes to ObjectDumping.Internal
