+++ ObjectDumper/Extensions/PropertyInfoExtensions.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Reflection;
 using System.Text;
 
-namespace ObjectDumperLib.Extensions
+namespace ObjectDumping.Extensions
 {
     internal static class PropertyInfoExtensions
     {
