+++ ObjectDumper/Extensions/StringExtensions.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace ObjectDumperLib.Extensions
+namespace ObjectDumping.Extensions
 {
     internal static class StringExtensions
     {
