+++ Tests/ObjectDumper.Tests/ObjectDumperTests.cs
@@ -3,10 +3,11 @@ using FluentAssertions;
 using Xunit;
 using Xunit.Abstractions;
 
-using ObjectDumperLib_Tests.Testdata;
-using ObjectDumperLib_Tests.Utils;
+using ObjectDumping;
+using ObjectDumper_Tests.Testdata;
+using ObjectDumper_Tests.Utils;
 
-namespace ObjectDumperLib_Tests
+namespace ObjectDumper_Tests
 {
     [Collection(TestCollections.CultureSpecific)]
     public class ObjectDumperTests
@@ -30,7 +31,7 @@ namespace ObjectDumperLib_Tests
             // Assert
             this.testOutputHelper.WriteLine(dump);
             dump.Should().NotBeNull();
-            dump.Should().Be("{ObjectDumperLib_Tests.Testdata.Person}\n\r  Name: \"Thomas\"\n\r  Char: \n\r  Age: 30\n\r  GetOnly: 11\n\r  Bool: False\n\r  Byte: 0\n\r  ByteArray: ...\n\r    1\n\r    2\n\r    3\n\r    4\n\r  SByte: 0\n\r  Float: 0\n\r  Uint: 0\n\r  Long: 0\n\r  ULong: 0\n\r  Short: 0\n\r  UShort: 0\n\r  Decimal: 0\n\r  Double: 0\n\r  DateTime: 01.01.0001 00:00:00\n\r  NullableDateTime: null\n\r  Enum: Unspecified\n\r");
+            dump.Should().Be("{ObjectDumper_Tests.Testdata.Person}\n\r  Name: \"Thomas\"\n\r  Char: \n\r  Age: 30\n\r  GetOnly: 11\n\r  Bool: False\n\r  Byte: 0\n\r  ByteArray: ...\n\r    1\n\r    2\n\r    3\n\r    4\n\r  SByte: 0\n\r  Float: 0\n\r  Uint: 0\n\r  Long: 0\n\r  ULong: 0\n\r  Short: 0\n\r  UShort: 0\n\r  Decimal: 0\n\r  Double: 0\n\r  DateTime: 01.01.0001 00:00:00\n\r  NullableDateTime: null\n\r  Enum: Unspecified\n\r");
         }
 
         [Fact]
