+++ Tests/ObjectDumper.Tests/Testdata/OrderPropertyTestObject.cs
@@ -1,4 +1,4 @@
-﻿namespace ObjectDumperLib_Tests.Testdata
+﻿namespace ObjectDumper_Tests.Testdata
 {
     public class OrderPropertyTestObject
     {
