+++ Tests/ObjectDumper.Tests/Testdata/Organization.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace ObjectDumperLib_Tests.Testdata
+namespace ObjectDumper_Tests.Testdata
 {
     public class Organization
     {
